cleanup and lint
