+++ project/assets/js/containers/Root.js
@@ -1,5 +1,23 @@
-if (process.env.NODE_ENV === 'production') {
-    module.exports = require('./Root.prod');
-} else {
-    module.exports = require('./Root.dev');
+import React, { Component, PropTypes } from 'react';
+import { Provider } from 'react-redux';
+import { ReduxRouter } from 'redux-router';
+import { loadConfig } from 'actions/Config';
+
+export default class Root extends Component {
+    componentWillMount() {
+        this.props.store.dispatch(loadConfig());
+    }
+
+    render() {
+        const { store } = this.props;
+        return (
+            <Provider store={store}>
+                    <ReduxRouter />
+            </Provider>
+        );
+    }
 }
+
+Root.propTypes = {
+    store: PropTypes.object.isRequired,
+};
