--- project/assets/js/containers/Root.prod.js
@@ -1,22 +1,0 @@
-import React, { Component, PropTypes } from 'react';
-import { Provider } from 'react-redux';
-import { ReduxRouter } from 'redux-router';
-
-export default class Root extends Component {
-    componentWillMount() {
-        store.dispatch(loadConfig());
-    }
-
-    render() {
-        const { store } = this.props;
-        return (
-            <Provider store={store}>
-                <ReduxRouter />
-            </Provider>
-        );
-    }
-}
-
-Root.propTypes = {
-    store: PropTypes.object.isRequired,
-};
