--- project/index.html
@@ -1,10 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>React Transform Boilerplate</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script src="/dist/bundle.js"></script>
-  </body>
-</html>
